custom-faster-r-cnn/requirements.txt
torch
torchvision
albumentations
pandas
numpy
opencv-python-headless
scikit-learn
matplotlib
seaborn
tqdm
wandb
pillow
